split out some app utils to make them deno friendly
